+++ CropGrowthAdjustments/Core/Integrations/IJsonAssetsApi.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace CropGrowthAdjustments.Core.Integrations;
+
+public interface IJsonAssetsApi
+{
+    event EventHandler ItemsRegistered;
+}