+++ CropGrowthAdjustments/Types/IJsonAssetsApi.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace CropGrowthAdjustments.Types
-{
-    public interface IJsonAssetsApi
-    {
-        event EventHandler ItemsRegistered;
-    }
-}